Code refactoring

+++ fact.h
@@ -5,4 +5,4 @@ int parse(const int argc, char* argv[]);
 
 void throw_error(const int argc); 
 
-unsigned long long factorial(unsigned int number);
+unsigned long long factorial(unsigned long long number);
